+++ server/conf/settings.py
@@ -36,6 +36,8 @@ SERVERNAME = "Druidia"
 
 # Typeclass for account objects (linked to a character) (fallback)
 BASE_ACCOUNT_TYPECLASS = "typeclasses.base.Account"
+# Typeclass for Channel (fallback).
+BASE_CHANNEL_TYPECLASS = "typeclasses.base.Channel"
 # Typeclass for Exit objects (fallback).
 BASE_EXIT_TYPECLASS = "typeclasses.base.Exit"
 # Typeclass for character objects linked to an account (fallback)
@@ -47,8 +49,6 @@ BASE_ROOM_TYPECLASS = "typeclasses.base.Room"
 # Typeclass for Scripts (fallback). You usually don't need to change this
 # but create custom variations of scripts on a per-case basis instead.
 BASE_SCRIPT_TYPECLASS = "typeclasses.base.Script"
-# Typeclass for Channel (fallback).
-#BASE_CHANNEL_TYPECLASS = "typeclasses.base.Channel"
 
 ######################################################################
 # Settings given in secret_settings.py override those in this file.
